Moved bulk metadata updating over to services-worker.  Added configuration for storage location for update files
