+++ metadata/src/main/java/edu/unc/lib/dl/util/RedisWorkerConstants.java
@@ -10,6 +10,7 @@ public class RedisWorkerConstants {
 	public static final String JOB_STATUS_PREFIX = "job-status:";
 	public static final String BULK_UPDATE_PREFIX = "bulk-update:";
 	public static final String BULK_RESUME_PREFIX = "bulk-resume:";
+	public static final String BULK_UPDATE_QUEUE = "bulk-md-update";
 
 	public static enum DepositField {
 		uuid, state, actionRequest, contactName, depositorName, intSenderIdentifier, intSenderDescription,
