+++ services-worker/src/main/resources/update-complete-html.txt
