+++ services-worker/src/main/resources/update-failed-html.txt
